candidates_section: extract nav into its own component
when many candidates, repeat nav before candidates in order to
empty the queue when cancelling for example
